start_dbsdk_directory
