+++ app/services/accounting.py
@@ -19,9 +19,9 @@ from app.schemas.callback import CallBack, HttpRequestCallBackConfig
 from app.schemas.task import TaskAccountingInfo, TaskDefinition
 from app.types import CallBackAction, CallBackEvent, TaskType
 from app.utils.http import make_http_request
+from obi_one.db_sdk.db_sdk import select_json_asset_content
 from obi_one.scientific.tasks.circuit_extraction.estimate import estimate_circuit_extraction_count
 from obi_one.scientific.tasks.skeletonization.estimate import estimate_skeletonization_count
-from obi_one.utils.db_sdk import select_json_asset_content
 
 CIRCUIT_SCALE_TO_SERVICE_SUBTYPE = {
     CircuitScale.single: ServiceSubtype.SINGLE_SIM,
