+++ app/services/task.py
@@ -30,7 +30,7 @@ from app.schemas.task import (
     TaskLaunchSubmit,
 )
 from app.types import CallBackAction, CallBackEvent, TaskType
-from obi_one.utils import db_sdk
+from obi_one.db_sdk import db_sdk
 
 
 def submit_task_job(
