+++ obi_one/scientific/tasks/skeletonization/estimate.py
@@ -11,9 +11,9 @@ from entitysdk import models
 from entitysdk.client import Client
 from entitysdk.types import AssetLabel, ContentType, FetchFileStrategy
 
+from obi_one.db_sdk.db_sdk import get_entity_asset_by_label
 from obi_one.scientific.from_id.em_cell_mesh_from_id import EMCellMeshFromID
 from obi_one.scientific.tasks.skeletonization.config import SkeletonizationSingleConfig
-from obi_one.utils.db_sdk import get_entity_asset_by_label
 
 
 def _compute_mesh_surface_area(db_client: Client, cell_mesh: EMCellMeshFromID) -> float:
